+++ react/typings/global.d.ts
@@ -57,3 +57,16 @@ interface UserDetails {
   email: string
   canImpersonate: boolean
 }
+
+interface CustomField {
+  name: string
+  type: 'text' | 'dropdown'
+  value?: string
+  dropdownValues?: Array<{ label: string; value: string }>
+  useOnRegistration?: boolean
+}
+
+interface CustomFieldSetting {
+  name: string
+  type: 'text' | 'dropdown'
+}